+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/general/avatar/group-avatar/group-avatar-users.tsx
