+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/general/notes/dialogs/note-information-dialog.tsx
@@ -4,9 +4,9 @@ import "~/sass/elements/form.scss";
 import "~/sass/elements/wizard.scss";
 import Button from "~/components/general/button";
 import NotesListItem, { NotesListItemProps } from "../notes-item-list-item";
-import GroupAvatarUsers from "~/components/general/avatar/subtypes/group-components/group-avatar-users";
+import GroupAvatarUsers from "~/components/general/avatar/group-avatar/group-avatar-users";
 import { useTranslation } from "react-i18next";
-import { AvatarGroupUser } from "~/components/general/avatar/subtypes/group-components/group-avatar-user";
+import { AvatarGroupUser } from "~/components/general/avatar/group-avatar/group-avatar-user";
 import { NoteReceiver } from "~/generated/client";
 /**
  * NoteInformationDialogProps
